refactor: data store
