+++ bluebox/__init__.py
@@ -56,7 +56,7 @@ from .utils.exceptions import (
 from . import agents
 from . import cdp
 from . import data_models
-from . import routine_discovery
+from . import llms
 from . import utils
 
 __all__ = [
@@ -87,7 +87,7 @@ __all__ = [
     "agents",
     "cdp",
     "data_models",
-    "routine_discovery",
+    "llms",
     "utils",
 ]
