+++ bluebox/sdk/discovery.py
@@ -21,7 +21,7 @@ from pydantic import BaseModel
 PACKAGE_ROOT = Path(__file__).resolve().parent.parent
 
 from ..agents.routine_discovery_agent import RoutineDiscoveryAgent
-from ..routine_discovery.data_store import LocalDiscoveryDataStore
+from ..llms.infra.data_store import LocalDiscoveryDataStore
 from ..data_models.routine.routine import Routine
 from ..data_models.routine_discovery.message import RoutineDiscoveryMessage
 from ..data_models.routine_discovery.llm_responses import TestParametersResponse
