+++ scripts/run_guide_agent.py
@@ -65,7 +65,7 @@ from bluebox.data_models.llms.interaction import (
 )
 from bluebox.data_models.routine.routine import Routine
 from bluebox.llms.tools.guide_agent_tools import validate_routine
-from bluebox.routine_discovery.data_store import DiscoveryDataStore, LocalDiscoveryDataStore
+from bluebox.llms.infra.data_store import DiscoveryDataStore, LocalDiscoveryDataStore
 from bluebox.data_models.routine_discovery.message import RoutineDiscoveryMessage, RoutineDiscoveryMessageType
 from bluebox.sdk import BrowserMonitor
 from bluebox.sdk.discovery import RoutineDiscovery
